requirements.txt
anthropic>=0.25.0
yfinance>=0.2.40
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0
tabulate>=0.9.0
